Updated .clang-tidy to include additional header filters for KeireHub components, modified .gitattributes to enforce line endings for various file types, and enhanced .gitignore to exclude specific documentation site directories. Improved README.md with clearer project descriptions and updated commands for building and running the project. Added new documentation files for Animation Rigging and Architecture, and created a dedicated Asset Browser guide. Refactored CI workflows to utilize a reusable quality gate for better maintainability. Updated CHANGELOG.md to reflect these changes and enhancements.

+++ KeireClient/Source/Editor/LightingPanel.cpp
@@ -179,8 +179,9 @@ namespace KeireEditor
         {
             if (const auto progress = m_Controller.LightingBakeProgress())
             {
-                const auto fraction =
-                    progress->Total == 0U ? 0.0F : static_cast<float>(progress->Completed) / progress->Total;
+                const auto fraction = progress->Total == 0U ? 0.0F
+                                                            : static_cast<float>(progress->Completed) /
+                                                                  static_cast<float>(progress->Total);
                 ui.ProgressBar(fraction, {0.0F, 18.0F}, progress->CurrentPath.string());
             }
             else
